Convert Domain User Entity fields to value objects

+++ code/usersregistration-api/rest/api-rest-server/src/main/java/com/company/signup/api/dto/UserDTO.java
@@ -36,11 +36,11 @@ public class UserDTO {
   @NotNull
   @DecimalMin(value = "1.00")
   @DecimalMax(value = "2.30")
-  Float height;
+  Double height;
 
   @NotNull
   @Min(40)
   @Max(150)
-  Float weight;
+  Double weight;
 
 }
